Add test case from Mathieu Dutour:
Dear Matthias,

there is a change in the result of "zsolve" between the latest version on the web page
and version 1.5 for the input file in attachment.

For version 1.5 I obtained 21 inhomogeneous solutions which is the correct number.
For version 1.5.3 I obtain just 1 which is not correct.

Is it a change of user interface or a bug ?

Best,

  Mathieu