configure.ac: drop AX_CC_MAXOPT and AX_CXX_MAXOPT macros
These two macros try to set the compiler to the highest level of
optimization, but they rely on probing the compiler vendor to work,
and were written in 2005. They leave several things to be desired:

  * They don't know about clang
  * They are missing several important flags (-march=native)
  * They clobber the user's own CFLAGS/CXXFLAGS

The downsides to this outweigh the benefits: it makes debugging a
hassle; and even for the purposes of optimization, one gains an -O3
but loses the ability to pass -march=native, -flto, etc.